Azure Functions: Publish using wrong dotnet version #661
